+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Ensure highlighting of unicode forall syntax matches that of alphabetic forall syntax
   ([#175](https://github.com/JustusAdam/language-haskell/issues/175)).
 - Fix highlighting of empty quasi-quotations such as `[i||]`.
+- Support GHC 9.0 pragmas: `LinearTypes`, `QualifiedDo`, `LexicalNegation`.
+- Add support for [modifiers](https://github.com/ghc-proposals/ghc-proposals/pull/370), e.g. `id :: a %1 -> a`.
 
 ## 3.3.0 - 25.06.2020